+++ src/app/core/models/comment.ts
@@ -1,0 +1,47 @@
+export interface IEventComment {
+  id: string;
+  eventId: string;
+  userUid: string;
+  userName: string;
+  userImage: string;
+  message: string;
+  timestamp: string;
+}
+
+export class EventComment implements IEventComment {
+
+  constructor(
+    public id: string,
+    public eventId: string,
+    public userUid: string,
+    public userName: string,
+    public userImage: string,
+    public message: string,
+    public timestamp: string
+  ) {
+  }
+}
+
+export interface INoticeComment {
+  id: string;
+  noticeId: string;
+  userUid: string;
+  userName: string;
+  userImage: string;
+  message: string;
+  timestamp: string;
+}
+
+export class NoticeComment implements INoticeComment {
+
+  constructor(
+    public id: string,
+    public noticeId: string,
+    public userUid: string,
+    public userName: string,
+    public userImage: string,
+    public message: string,
+    public timestamp: string
+  ) {
+  }
+}