+++ src/app/core/services/comments.service.ts
@@ -9,18 +9,21 @@ import { AngularFireAuth } from '@angular/fire/auth';
 
 import { Observable } from 'rxjs';
 
-import { IEventComment } from '@models/event-comment';
+import { IEventComment, INoticeComment } from '@models/comment';
 import { AppointmentsService } from '@services/appointments.service';
 
 const EVENTS_COMMENTS_COLLECTION = 'eventos-comentarios';
+const NOTICES_COMMENTS_COLLECTION = 'avisos-comentarios';
 
 @Injectable({
   providedIn: 'root'
 })
-export class EventsCommentsService {
+export class CommentsService {
 
   private eventCommentsCollection!: AngularFirestoreCollection<IEventComment>;
   private eventCommentsDoc!: AngularFirestoreDocument<IEventComment>;
+  private noticeCommentsCollection!: AngularFirestoreCollection<INoticeComment>;
+  private noticeCommentsDoc!: AngularFirestoreDocument<INoticeComment>;
 
   constructor(
     private afAuth: AngularFireAuth,
@@ -28,6 +31,7 @@ export class EventsCommentsService {
     private appointmentSrv: AppointmentsService,
   ) {
     this.eventCommentsCollection = afs.collection(EVENTS_COMMENTS_COLLECTION);
+    this.noticeCommentsCollection = afs.collection(NOTICES_COMMENTS_COLLECTION);
   }
 
   public getAllEventComments(eventId: string): Observable<IEventComment[]> {
@@ -39,6 +43,17 @@ export class EventsCommentsService {
     );
 
     return this.eventCommentsCollection.valueChanges();
+  }
+
+  public getAllNoticeComments(noticeId: string): Observable<INoticeComment[]> {
+
+    this.noticeCommentsCollection = this.afs.collection<INoticeComment>(
+      NOTICES_COMMENTS_COLLECTION,
+      ref => ref.where('noticeId', '==', noticeId)
+                .orderBy('timestamp', 'desc')
+    );
+
+    return this.noticeCommentsCollection.valueChanges();
   }
 
   async addEventComment(eventId: string, message: string): Promise<any> {
@@ -58,8 +73,30 @@ export class EventsCommentsService {
     });
   }
 
+  async addNoticeComment(noticeId: string, message: string): Promise<any> {
+
+    const currentUser = await this.afAuth.currentUser;
+    const id: string = this.afs.createId();
+    const timestamp = this.appointmentSrv.getTimestamp();
+
+    this.noticeCommentsCollection.doc(id).set({
+      id,
+      noticeId,
+      timestamp,
+      userUid: currentUser.uid,
+      userName: currentUser.displayName,
+      userImage: currentUser.photoURL,
+      message,
+    });
+  }
+
   deleteEventComment(id: string): void {
     this.eventCommentsDoc = this.afs.doc<IEventComment>(`${EVENTS_COMMENTS_COLLECTION}/${id}`);
     this.eventCommentsDoc.delete();
+  }
+
+  deleteNoticeComment(id: string): void {
+    this.noticeCommentsDoc = this.afs.doc<INoticeComment>(`${NOTICES_COMMENTS_COLLECTION}/${id}`);
+    this.noticeCommentsDoc.delete();
   }
 }