+++ src/app/features/notices/notices.module.ts
@@ -10,6 +10,8 @@ import { NoticeEditComponent } from '@features/notices/notice-edit/notice-edit.c
 import { NoticesDashboardComponent } from '@features/notices/notices-dashboard/notices-dashboard.component';
 import { NoticeItemDetailComponent } from '@features/notices/notice-item-detail/notice-item-detail.component';
 import { NoticeExpansionPanelComponent } from '@features/notices/notice-expansion-panel/notice-expansion-panel.component';
+import { NoticeCommentsDialogComponent } from '@features/notices/notice-comments-dialog/notice-comments-dialog.component';
+
 
 @NgModule({
   declarations: [
@@ -19,6 +21,7 @@ import { NoticeExpansionPanelComponent } from '@features/notices/notice-expansio
     NoticesDashboardComponent,
     NoticeExpansionPanelComponent,
     NoticeItemDetailComponent,
+    NoticeCommentsDialogComponent,
   ],
   imports: [
     HttpClientModule,